[model/API] Sync model cc API with c API
Sync the cc-API of model with the c-API
Some unnecessary functions have been removed.
Some others have been renamed.

**Self evaluation:**
1. Build test: [x]Passed [ ]Failed [ ]Skipped
2. Run test: [x]Passed [ ]Failed [ ]Skipped
